requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
banks==2.1.2
beautifulsoup4==4.13.4
bibtexparser==1.4.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
e2b==1.5.2
e2b-code-interpreter==1.5.1
exceptiongroup==1.3.0
fastmcp==2.8.1
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
llama-cloud==0.1.26
llama-cloud-services==0.6.34
llama-index==0.12.43
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.43
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-instrumentation==0.1.0
llama-index-llms-openai==0.4.7
llama-index-llms-openai-like==0.4.0
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-tools-mcp==0.2.5
llama-index-workflows==0.2.2
llama-parse==0.6.34
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.4
mdurl==0.1.2
multidict==6.5.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
openai==1.88.0
openapi-pydantic==0.5.1
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.47.0
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1